Add initial Secret class for secret management.
At the moment, we're limiting secret management in Grouper to a frontend role only;
all distribution, storage, rotation, and other concerns are handled by a plugin or
another system which interacts with Grouper via a plugin. In order to handle as wide
an array of secret management and distribution platforms as possible, the interface
through which Grouper interacts with the plugins is intentionally generic.