Add all secrets view and create secrets handlers
Also introduces a new fe util function called paginate_results to
remove duplicated code required to paginate long lists. Adds the
form_field macro to macros/ui.html to prevent future forms from
having to duplicate that macro as well. Additionally a form_attr
attribute (with type str) has been added to the Secret class. It's
intended use is as additional information to the form attribute;
for instance, the number of bytes for an automatically generated
secret, or the secret itself for manually created secrets.