Add ability to edit secrets.
Also rename the SecretCreateForm to SecretForm because it is used for
both creating and editing secrets. SecretsView and SecretView now
share a substantial amount of code, mainly code dealing with parsing
and presenting the SecretForm. Editing the name of secrets is not
allowed because names are how secrets are idenitifed.