Change secret delete to use POST
The dev environment has poor support for DELETE requests and HTML
doesn't support DELETE as well as POST. Added a decorator and macro
for submitting forms and treating the submission as if it was sent
as that http verb so we can easily switch over to using the proper
http verbs if they ever get proper html support. Also fix a couple
bugs; one where the distribution was improprely getting converted
for the edit form, and the other where we didn't have a link to the
secret detail page on the secrets view page.