Add database session as argument to secret commit/delete/get
This allows plugins to rely on the database (by storing the secrets
in the database or otherwise). This is especially required because
plugins must return secrets which have a proper Group object set
as the owner, which must be obtained from the database. This also
adds to_dict and from_dict methods to the Secret class for use when
needing to convert the secrets to or from a JSON serializable format.