Add proper support for Secret subclasses
Removed the form and form_attr fields from the Secret class.
Instead of using those fields, we now allow plugins to create
subclasses of the Secret class and expose that to the main
codebase. These subclasses can have arbitrary additional fields,
and include a form so that each secret type supported by plugins
has a unique form tailored to it's particular use case.

This change was motivated by the difficulty and poor UI experience
encountered when trying to support a wide range of various secret
types, which often have very little in common other than the fact
that they are sensitive and should only be distributed to a small
subset of machines.

In order to facilitate the use of arbitrary forms, several new macros
were added; the form macro takes in a WTForms Form object and displays
all of the fields without having to manually specify each field (as
was the case previously). The form_modal macro was also added to allow
for cleanly supporting arbitrary numbers of form modals in a single page.